pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dense-sparse-extractor"
version = "0.0.0"
description = "Dense/sparse feature extraction experiments."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "PyYAML",
  "tqdm",
  "matplotlib",
  "ipykernel",
]

[tool.setuptools]
packages = ["dense_sparse_extractor"]

